backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
groq==0.5.0
supabase==2.4.5
pinecone-client==3.2.2
psycopg2-binary
PyJWT==2.8.0